refactor: move cache to pkg, simplify handlers (#42)
* refactor: move cache to pkg

* refactor: simplify handlers
