+++ internal/files/store/file_test.go
@@ -10,8 +10,8 @@ import (
 	"testing"
 
 	"github.com/azure/peerd/internal/files"
-	"github.com/azure/peerd/internal/files/cache"
 	remotetests "github.com/azure/peerd/internal/remote/tests"
+	"github.com/azure/peerd/pkg/cache"
 	"github.com/azure/peerd/pkg/discovery/routing/tests"
 )
