+++ internal/handlers/files/handler_test.go
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-package handlers
+package files
 
 import (
 	"context"
