+++ internal/handlers/v2/handler_test.go
@@ -1,4 +1,4 @@
-package handlers
+package v2
 
 import (
 	"context"
