+++ internal/handlers/v2/registry.go
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-package oci
+package v2
 
 import (
 	"fmt"
